+++ R/suitability.R
@@ -23,6 +23,31 @@ g3_suitability_andersen <- function (p0, p1, p2, p3 = p4, p4, p5 = ~pred_stock__
                  p5 = p5))
 }
 
+g3_suitability_andersenfleet <- function (
+        p0 = g3_parameterized('andersen.p0', by_stock = by_stock, value = 0,
+                              optimise = FALSE),
+        p1 = g3_parameterized('andersen.p1', by_stock = by_stock, value = log(2)),
+        p2 = g3_parameterized('andersen.p2', by_stock = by_stock, value = 1,
+                              optimise = FALSE),
+        p3 = g3_parameterized('andersen.p3', by_stock = by_stock, value = 0.1,
+                              exponentiate = exponentiate),
+        p4 = g3_parameterized('andersen.p4', by_stock = by_stock, value = 0.1,
+                              exponentiate = exponentiate),
+        p5 = quote( stock__maxmidlen ),
+        by_stock = TRUE,
+        exponentiate = TRUE) {
+    f_substitute(~p0 +
+        p2 * exp(-(log(p5/stock__midlen) - p1)**2/p3) *
+        bounded_vec(1000*(p1 - log(p5/stock__midlen)),0,1) +
+        p2 * exp(-(log(p5/stock__midlen) - p1)**2/p4) *
+        bounded_vec(1000*(log(p5/stock__midlen) - p1),0,1), list(
+            p0 = p0,
+            p1 = p1,
+            p2 = p2,
+            p3 = p3,
+            p4 = p4,
+            p5 = p5))
+}
 
 g3_suitability_gamma <- function(alpha, beta, gamma){
   ## I'm not sure why beta and gamma are not just a single parameter but